# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestão de livros",
        "title": "Book Store API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Realizar login do usuário",
                "parameters": [
                    {
                        "description": "Dados para login do usuário",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UserLoginPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tokens de acesso e refresh",
                        "schema": {
                            "$ref": "#/definitions/types.UserLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Validation errors for payload",
                        "schema": {
                            "$ref": "#/definitions/types.BadRequestStructResponse"
                        }
                    },
                    "404": {
                        "description": "No user found with the given email",
                        "schema": {
                            "$ref": "#/definitions/types.NotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/types.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Atualizar tokens (Refresh Token)",
                "parameters": [
                    {
                        "description": "Payload contendo o refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RefreshTokenPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Novos tokens de acesso e refresh",
                        "schema": {
                            "$ref": "#/definitions/types.UpdateRefreshTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Validation errors for payload",
                        "schema": {
                            "$ref": "#/definitions/types.BadRequestStructResponse"
                        }
                    },
                    "401": {
                        "description": "Refresh token is invalid or has been expired",
                        "schema": {
                            "$ref": "#/definitions/types.UnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "No refresh token found with the given user ID",
                        "schema": {
                            "$ref": "#/definitions/types.NotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "An unexpected error occurred",
                        "schema": {
                            "$ref": "#/definitions/types.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.BadRequestStructResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "types.InternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "types.NotFoundResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "types.RefreshTokenPayload": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "types.UnauthorizedResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "types.UpdateRefreshTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "types.UserLoginPayload": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "types.UserLoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}